servicio-auditoria/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
ItsDangerous==2.1.2
MarkupSafe==2.1.3
click==8.1.7
psycopg2-binary==2.9.7
